Исправил ошибку, user_id BigInteger

+++ db_init.py
@@ -1,4 +1,4 @@
-from sqlalchemy import create_engine, Column, Integer, String, Boolean, ForeignKey
+from sqlalchemy import create_engine, Column, BigInteger, String, Boolean, ForeignKey
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship, sessionmaker
 import pandas as pd
@@ -9,7 +9,7 @@ Base = declarative_base()
 class UserSection(Base):
     __tablename__ = 'user_sections'
     __table_args__ = {'extend_existing': True}
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger ,primary_key=True)
     section_name = Column(String(255), nullable=False)
     user_associations = relationship("UserSectionAssociation", back_populates="user_section", cascade="all, delete-orphan")
 
@@ -17,7 +17,7 @@ class UserSection(Base):
 class User(Base):
     __tablename__ = 'users'
     __table_args__ = {'extend_existing': True}
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     first_name = Column(String(255))
     is_work = Column(Boolean, default=True)
     username = Column(String)
@@ -31,8 +31,8 @@ class SkuToSection(Base):
 
 class UserSectionAssociation(Base):
     __tablename__ = 'user_section_association'
-    user_id = Column(Integer, ForeignKey('users.user_id', ondelete='CASCADE'), primary_key=True)
-    user_section_id = Column(Integer, ForeignKey('user_sections.id',ondelete='CASCADE'), primary_key=True)
+    user_id = Column(BigInteger, ForeignKey('users.user_id', ondelete='CASCADE'), primary_key=True)
+    user_section_id = Column(BigInteger, ForeignKey('user_sections.id',ondelete='CASCADE'), primary_key=True)
     # Создание двусторонних связей с User и UserSection
     user = relationship("User", back_populates="user_sections")
     user_section = relationship("UserSection", back_populates="user_associations")
